requirements.txt
paho-mqtt
python-dotenv
opencv-contrib-python
vidgear
adafruit-circuitpython-ina260
python-dotenv